Use dynamic buffer instead of fixed size buffer
Use vector<char> as a dynamic buffer with an initial buffer size of 10000 which should be big enough for most cases. When the buffer size is not big enough, resize the buffer with the min buffer size set in olen out param by the mbedtls_pem_write_buffer function and retry again.