CBL-4237 : Pass cert chain to the root_cert_locator() (#12)
* CBL-4237 : Pass cert chain to the root_cert_locator()

Background for CBL-4234:

In LiteCore PublicKey+Apple.mm’s findSigningRootCert() for TLS CA cert validation, a SecTrust is created with only a leaf cert without its intermediate certs. As a result, when calling SecTrustEvaluateWithError() to evaluate the SecTrust (a technique to get all certs including the root CA cert used for verifying a CA signed cert), there is an error logged as follows:

```
[TLS] error: SecTrustEvaluateWithError failed: Error Domain=NSOSStatusErrorDomain Code=-25318 "errKCCreateChainFailed / errSecCreateChainFailed:  / The attempt to create a certificate chain failed."
```

Fix:

This commit is constructing and passing the cert chain (in PEM format) to the root_cert_locator(). The passed cert chain, will eventually be passed to the findSigningRootCert() that Apple platform can use when constructing an array of SecCertRef for creating a SecTrust. Note that there will be a corresponding fix in LiteCore regarding this as well.

* Change max buffer size to 10000 to match LiteCore’s mbedUtils.cc and use dynamic buffer instead of fixed size buffer

Use vector<char> as a dynamic buffer with an initial buffer size of 10000 which should be big enough for most cases. When the buffer size is not big enough, resize the buffer with the min buffer size set in olen out param by the mbedtls_pem_write_buffer function and retry again.